pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "heartbeat-notify"
version = "0.1.0"
description = "A tool to monitor file updates and notify via Discord webhook if stalled."
readme = "README.md"
authors = [
  { name = "User", email = "user@example.com" },
]
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests",
    "pyyaml",
    "click",
    "pydantic",
]

[project.scripts]
heartbeat-notify = "heartbeat_notify.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/heartbeat_notify"]
